net: introduce global.net.server
The server to use is independent of the network device, there is
not much point to make the server specific to a network device.

This introduces global.net.server as the serverip which is
used as standard NFS/tftp server. The previously used eth
device specific parameters still exist, but are only aliases
for the global single variable.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>