Pass barebox version to kernel
When userspace is interested in the barebox version it has currently no
way of reliably reading it. Add it to the kernel command line as it's an
established way to pass information from the bootloader to the kernel.
If CONFIG_FLEXIBLE_BOOTARGS is enabled then the barebox version is
passed in the "bootloader.version=" variable.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>