net: fec_imx: skip fec_halt if the FEC wasn't started
If the FEC driver is enabled, but the ethernet device was never
opened due to booting from another source the wait for graceful
transmit stop will always fail. This introduces an unnecessary
boot delay and prints a distracting warning.

As there is no point in trying to shut down the FEC if it was
never started, just skip all of fec_halt in that case.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>